refactor: move FEE_CONFIG_KEYS to constants
