+++ src/components/settings/FeeConfigModal.tsx
@@ -5,13 +5,13 @@ import { Formik, FormikErrors, FormikProps, Field } from 'formik'
 import classNames from 'classnames'
 import Sprite from '../Sprite'
 import { TxFeeInputField, validateTxFee } from './TxFeeInputField'
-import { FEE_CONFIG_KEYS, FeeValues, useLoadFeeConfigValues } from '../../hooks/Fees'
+import { FeeValues, useLoadFeeConfigValues } from '../../hooks/Fees'
 import { useUpdateConfigValues } from '../../context/ServiceConfigContext'
 import { isDebugFeatureEnabled } from '../../constants/debugFeatures'
+import { FEE_CONFIG_KEYS, JM_MAX_SWEEP_FEE_CHANGE_DEFAULT } from '../../constants/config'
 import ToggleSwitch from '../ToggleSwitch'
 import { isValidNumber, factorToPercentage, percentageToFactor } from '../../utils'
 import BitcoinAmountInput, { AmountValue, toAmountValue } from '../BitcoinAmountInput'
-import { JM_MAX_SWEEP_FEE_CHANGE_DEFAULT } from '../../constants/config'
 import styles from './FeeConfigModal.module.css'
 
 const __dev_allowFeeValuesReset = isDebugFeatureEnabled('allowFeeValuesReset')
